rename parseCli to parse-compiler-opts
