+++ stopify/src/compile.ts
@@ -1,7 +1,7 @@
 import 'source-map-support/register'
 import * as fs from 'fs-extra';
 import { stopify } from './index';
-import * as parseCli from 'stopify-continuations/dist/src/compiler/parseCli';
+import * as parseCli from 'stopify-continuations/dist/src/compiler/parse-compiler-opts';
 
 stopify(parseCli.srcPath, parseCli.compilerOpts)
   .then(dstCode => {
